Exit the server automatically if no instance exists, replace the sample class
