+++ src/Files.App.Server/Files.App.Server.csproj
@@ -50,4 +50,8 @@
         </None>
     </ItemGroup>
 
+    <ItemGroup>
+        <ProjectReference Include="..\Files.Shared\Files.Shared.csproj" />
+    </ItemGroup>
+
 </Project>
